Don't install libsdl1.2-dev
This package is only needed because there is a bug in the SDL_gfx build script in the psplibraries repository. This is addressed in https://github.com/pspdev/psplibraries/pull/43. When that PR has been merged, the entire psptoolchain can build without  libsdl1.2-dev installed. No including it should shave at least 100 megabytes, probably more, off of the docker container.